change ChannelOptions limit from usize to u64

+++ cable/src/lib.rs
@@ -32,7 +32,7 @@ pub struct ChannelOptions {
     pub channel: Channel,
     pub time_start: Timestamp,
     pub time_end: Timestamp,
-    pub limit: usize,
+    pub limit: u64,
 }
 
 #[derive(Clone, Debug, PartialEq)]
